requirements.txt
django_auth_ldap-1.1.4
python-2.7.10
djangorestframework-3.3.2
django-admin-bootstrapped-2.1.0
django-auth-ldap-1.2.1
django-bootstrap-tags-1.2.0
djangorestframework-3.3.0
Django-1.8.6
djangorestframework-jsonp
requests
uWSGI
